+++ loinc2hpo-core/src/main/java/org/monarchinitiative/loinc2hpo/ResourceCollection.java
@@ -3,7 +3,7 @@ package org.monarchinitiative.loinc2hpo;
 import com.google.common.collect.ImmutableMap;
 import org.monarchinitiative.loinc2hpo.exception.MalformedLoincCodeException;
 import org.monarchinitiative.loinc2hpo.exception.UnrecognizedLoincCodeException;
-import org.monarchinitiative.loinc2hpo.io.HPOParser;
+import org.monarchinitiative.loinc2hpo.io.HpoOntologyParser;
 import org.monarchinitiative.loinc2hpo.io.LoincAnnotationSerializationFactory;
 import org.monarchinitiative.loinc2hpo.loinc.LOINC2HpoAnnotationImpl;
 import org.monarchinitiative.loinc2hpo.loinc.LoincEntry;
@@ -18,6 +18,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -37,6 +38,7 @@ public class ResourceCollection {
     private String annotationMapPath;
     private String loincPanelPath;
     //private String loincPanelAnnotationPath;
+    private HpoOntologyParser hpoOntologyParser;
     private HpoOntology hpo;
     private Map<TermId, Term> termidTermMap;
     private Map<String, Term> termnameTermMap;
@@ -91,47 +93,32 @@ public class ResourceCollection {
         return this.loincIdSet;
     }
 
-    public Map<TermId, Term> hpoTermMap() {
+    private void parseHPO() throws PhenolException, FileNotFoundException {
+        if (this.hpoOboPath != null) {
+            hpoOntologyParser = new HpoOntologyParser(this.hpoOboPath);
+        } else if (this.hpoOwlPath != null) {
+            hpoOntologyParser = new HpoOntologyParser(this.hpoOwlPath);
+        }
+        hpoOntologyParser.parseOntology();
+    }
+
+    public Map<TermId, Term> hpoTermMap() throws PhenolException, FileNotFoundException {
         if (this.termidTermMap != null) {
             return this.termidTermMap;
         }
 
-        HpOboParser hpoOboParser = new HpOboParser(new File(this.hpoOboPath));
-        HpoOntology hpo = null;
-        try {
-            hpo = hpoOboParser.parse();
-        } catch (PhenolException e) {
-            e.printStackTrace();
-        }
-        ImmutableMap.Builder<TermId,Term> termmap = new ImmutableMap.Builder<>();
-        if (hpo !=null) {
-            List<Term> res = hpo.getTermMap().values().stream().distinct()
-                    .collect(Collectors.toList());
-            res.forEach( term -> termmap.put(term.getId(),term));
-        }
-        this.termidTermMap = termmap.build();
-        return this.termidTermMap;
+        parseHPO();
+
+        return hpoOntologyParser.getTermMap2();
     }
 
-    public Map<String, Term> hpoTermMapFromName() {
-        if (this.termnameTermMap != null) {
+    public Map<String, Term> hpoTermMapFromName() throws PhenolException, FileNotFoundException {
+        if (this.termnameTermMap != null){
             return this.termnameTermMap;
         }
 
-        HpOboParser hpoOboParser = new HpOboParser(new File(this.hpoOboPath));
-        HpoOntology hpo = null;
-        try {
-            hpo = hpoOboParser.parse();
-        } catch (PhenolException e) {
-            e.printStackTrace();
-        }
-        ImmutableMap.Builder<String,Term> termmap = new ImmutableMap.Builder<>();
-        if (hpo !=null) {
-            List<Term> res = hpo.getTermMap().values().stream().distinct()
-                    .collect(Collectors.toList());
-            res.forEach( term -> termmap.put(term.getName(),term));
-        }
-        this.termnameTermMap = termmap.build();
+        parseHPO();
+
         return this.termnameTermMap;
     }
 
@@ -156,9 +143,13 @@ public class ResourceCollection {
 
     }
 
-    public HpoOntology getHPO() throws PhenolException {
-        HpOboParser hpOboParser = new HpOboParser(new File(this.hpoOboPath));
-        this.hpo = hpOboParser.parse();
+    public HpoOntology getHPO() throws PhenolException, FileNotFoundException {
+        if (this.hpo != null) {
+            return this.hpo;
+        }
+
+        parseHPO();
+
         return this.hpo;
     }