+++ loinc2hpo-core/src/test/java/org/monarchinitiative/loinc2hpo/loinc/LOINC2HpoAnnotationImplTest.java
@@ -39,10 +39,9 @@ public class LOINC2HpoAnnotationImplTest {
 
 
     @BeforeClass
-    public static void setUp() {
+    public static void setUp() throws Exception {
         resourceCollection = new ResourceCollection();
         resourceCollection.setHpoOboPath(LOINC2HpoAnnotationImplTest.class.getResource("/obo/hp.obo").getPath());
-
         hpoTermMap = resourceCollection.hpoTermMapFromName();
     }